Split ExecutionBlockImpl
