+++ arangod/Aql/DistributeExecutor.h
@@ -25,7 +25,7 @@
 
 #include "Aql/BlocksWithClients.h"
 #include "Aql/DistributeClientBlock.h"
-#include "Aql/ExecutionBlockImpl.h"
+#include "Aql/ExecutionBlockImpl/ExecutionBlockImpl.h"
 #include "Aql/RegisterInfos.h"
 #include "Basics/ResultT.h"
