+++ arangod/Aql/IResearchViewNode.cpp
@@ -30,7 +30,7 @@
 #include "Aql/Collection.h"
 #include "Aql/Condition.h"
 #include "Aql/EmptyExecutorInfos.h"
-#include "Aql/ExecutionBlockImpl.h"
+#include "Aql/ExecutionBlockImpl/ExecutionBlockImpl.h"
 #include "Aql/ExecutionEngine.h"
 #include "Aql/ExecutionPlan.h"
 #include "Aql/IResearchViewExecutor.h"
