+++ arangod/Aql/MutexExecutor.h
@@ -25,7 +25,7 @@
 
 #include "Aql/BlocksWithClients.h"
 #include "Aql/DistributeClientBlock.h"
-#include "Aql/ExecutionBlockImpl.h"
+#include "Aql/ExecutionBlockImpl/ExecutionBlockImpl.h"
 #include "Aql/ExecutionState.h"
 #include "Aql/RegisterInfos.h"
 #include "Aql/SingleRowFetcher.h"
