+++ arangod/Aql/RemoteExecutor.h
@@ -25,7 +25,7 @@
 
 #include "Aql/AqlExecuteResult.h"
 #include "Aql/ClusterNodes.h"
-#include "Aql/ExecutionBlockImpl.h"
+#include "Aql/ExecutionBlockImpl/ExecutionBlockImpl.h"
 #include "Aql/RegisterInfos.h"
 
 #include <fuerte/message.h>
