+++ arangod/Aql/SortExecutor.cpp
@@ -23,7 +23,7 @@
 
 #include "SortExecutor.h"
 
-#include "Aql/ExecutionBlockImpl.h"
+#include "Aql/ExecutionBlockImpl/ExecutionBlockImpl.h"
 #include "Aql/InputAqlItemRow.h"
 #include "Aql/OutputAqlItemRow.h"
 #include "Aql/SingleRowFetcher.h"
