+++ arangod/Aql/SortedCollectExecutor.h
@@ -29,7 +29,7 @@
 #include "Aql/Aggregator.h"
 #include "Aql/AqlValueGroup.h"
 #include "Aql/ExecutionBlock.h"
-#include "Aql/ExecutionBlockImpl.h"
+#include "Aql/ExecutionBlockImpl/ExecutionBlockImpl.h"
 #include "Aql/ExecutionNode.h"
 #include "Aql/ExecutionState.h"
 #include "Aql/LimitStats.h"
