+++ arangod/CMakeLists.txt
@@ -359,7 +359,7 @@ set(LIB_ARANGO_AQL_SOURCES
   Aql/EnumerateListExecutor.cpp
   Aql/EnumeratePathsExecutor.cpp
   Aql/ExecutionBlock.cpp
-  Aql/ExecutionBlockImpl.cpp
+  Aql/ExecutionBlockImpl/Instances/Instances.cpp
   Aql/ExecutionEngine.cpp
   Aql/ExecutionNode.cpp
   Aql/ExecutionNodeId.cpp
